+++ NEWS.md
@@ -1,3 +1,7 @@
+* Added support for linkouts in metaFeatures table. See
+`?addMetaFeaturesLinkouts` for details. These linkouts will be displayed as
+favicons in the metaFeatures table in a future release of the app.
+
 * The "add" functions have a new argument, `reset`. If `reset = TRUE`, then the
 existing data is removed prior to adding the new data. The default continues to
 be to add to or modify any existing data.